refactor view for use in new engine version
