+++ lib/state.js
@@ -5,39 +5,44 @@
  * @param {object} The error object.
  * @param {object} The data object.
 */
-exports.state = function activateState (stream, options) {
+module.exports = function activateState (options, data, next) {
+
+    var state = (this._config.states || {})[options.name];
+    var template = (this._config.templates || {})[options.tmpl];
+    
+    if (!state) {
+        return next(new Error('View.state: State "' + options.name + '" not found.'));
+    }
 
     // check options
-    var template = this.t[options.tmpl];
-    if (!template || !options.name || !this.s[options.name]) {
-        return;
+    if (!template) {
+        return next(new Error('View.state: Template "' + option.tmpl + '"not found.'));
     }
 
     // activate state elements
-    for (var i = 0, state, selector, element; i < this.s[options.name].length; ++i) {
-        state = this.s[options.name][i];
+    state.forEach(function (config) {
 
         // call other states
         if (state.states) {
-            for (var ii = 0, l = state.states.length; ii < l; ++ii) {
-                activateState.call(this, {name: state.states[ii]});
-            }
+            state.states.forEach(function (stateName) {
+                activateState.call(this, {name: stateName});
+            });
         }
 
-        element = template.elements[options.element || state.element];
-        selector = state.sel || options.selector;
+        var element = template.elements[options.element || state.element];
+        var selector = state.sel || options.selector;
 
         // get dynamic or static selector
         if (!element && !selector) {
-
-            // retrun if no selector is found
-            return;
+            return next(new Error('View.state "' + options.name + '": DOM target not found.'));
         }
 
         // manipulate classes
         selector = typeof selector !== 'string' ? selector : (template.to || document).querySelectorAll(selector);
         manipulateClasses(element || selector, state);
-    }
+    });
+
+    next(null, data);
 };
 
 /**
